[fix]: changed some values from int to double

+++ src/main/java/clashclass/resources/ResourceManager.java
@@ -9,19 +9,19 @@ public interface ResourceManager {
      *
      * @param value the amount of resource to increase
      */
-    void increase (int value);
+    void increase (double value);
 
     /**
      * Decrease the amount of resource by a given value.
      *
      * @param value the amount of resource to decrease
      */
-    void decrease (int value);
+    void decrease (double value);
 
     /**
      * Returns the current value of the resource.
      *
      * @return the resource value as an int
      */
-    int getCurrentValue();
+    double getCurrentValue();
 }
